More changes.
Removed references to ClampMovement command and instead used InstantCommand with the clamp subsystem. will reimplement if this doesn't work.

Added enableTestBindings to the RobotContainer constructor to enable trent's drive test bindings. 

removed the references to the ScoreTwo auto routine. It will not work in competition i just can't bring myself to actually remove it.

Removed ManualArmExtension command. This will be handled with InstantCommand because its silly to do it in it's own subclass.